requirements.txt
aiohttp>=3.9.0
anthropic>=0.40.0
pyyaml>=6.0
markdown>=3.5
